move getAttendanceForPrisoner to IntegrationApiController
