+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsactivitiesmanagementapi/resource/AttendanceController.kt
@@ -216,71 +216,6 @@ class AttendanceController(private val attendancesService: AttendancesService) {
     reason = reason,
     categories = categories?.map { it.name },
   )
-
-  @GetMapping(value = ["/prisoner/{prisonerNumber}"])
-  @ResponseBody
-  @Operation(
-    summary = "Gets a list of prisoner attendance activities for a given date range",
-  )
-  @ApiResponses(
-    value = [
-      ApiResponse(
-        responseCode = "200",
-        description = "Successful call",
-        content = [
-          Content(
-            mediaType = "application/json",
-            array = ArraySchema(schema = Schema(implementation = Attendance::class)),
-          ),
-        ],
-      ),
-      ApiResponse(
-        responseCode = "400",
-        description = "Invalid request",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "401",
-        description = "Unauthorised, requires a valid Oauth2 token",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "403",
-        description = "Forbidden, requires an appropriate role",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "404",
-        description = "The attendance was not found.",
-        content = [
-          Content(
-            mediaType = "application/json",
-            schema = Schema(implementation = ErrorResponse::class),
-          ),
-        ],
-      ),
-    ],
-  )
-  @PreAuthorize("hasRole('ACTIVITIES__HMPPS_INTEGRATION_API')")
-  fun getAttendanceForPrisoner(
-    @PathVariable("prisonerNumber")
-    @Parameter(description = "Prisoner Number")
-    prisonerNumber: String,
-    @RequestParam("startDate", required = true)
-    @DateTimeFormat(iso = DateTimeFormat.ISO.DATE)
-    @Parameter(description = "Start date of query (required). Format YYYY-MM-DD.")
-    startDate: LocalDate,
-    @RequestParam("endDate", required = true)
-    @DateTimeFormat(iso = DateTimeFormat.ISO.DATE)
-    @Parameter(description = "End date of query (required). Format YYYY-MM-DD.")
-    endDate: LocalDate,
-    @RequestParam("prisonCode", required = false)
-    @Parameter(description = "The 3-character prison code.")
-    prisonCode: String? = null,
-  ): List<Attendance> = attendancesService.getPrisonerAttendance(
-    prisonerNumber = prisonerNumber,
-    startDate = startDate,
-    endDate = endDate,
-    prisonCode = prisonCode,
-  )
 }
+
+
